Some Initial Files/Button
I created some temporary files for now. I didnt know where in the repo we specifically wanted these but they can be moved around wherever. I made the on/off switch for the extension. The code can be moved to the main file when we all get our parts together.